Modularize Transitions View Model by outsourcing Bitmap related components
