+++ TgaBuilderWpfUi/App.DI.xaml.cs
@@ -278,10 +278,17 @@ namespace TgaBuilderWpfUi
 
                 presenter: GetBitmapFromFactory(sp, 2 * PANEL_WIDTH_INIT, PANEL_HEIGHT_INIT, true)));
 
+            services.AddTransient(sp => new TransitionsPresentersViewModel(
+                mediaFactory: sp.GetRequiredService<IMediaFactory>(),
+                transitionHelper: sp.GetRequiredService<ITransitionHelper>(),
+                bitmapOperations: sp.GetRequiredService<IBitmapOperations>()));
+
+
             services.AddTransient(sp => new TransitionViewModel(
                 mediaFactory: sp.GetRequiredService<IMediaFactory>(),
                 transitionHelper: sp.GetRequiredService<ITransitionHelper>(),
                 bitmapOperations: sp.GetRequiredService<IBitmapOperations>(),
+                transitionsPresentersViewModel: sp.GetRequiredService<TransitionsPresentersViewModel>(),
                 mainViewModel: sp.GetRequiredService<MainViewModel>()));
 
             services.AddTransient(sp => new ModificationsViewModel(
