+++ TgaBuilderWpfUi/View/TransitionWindowManualUserControl.xaml
@@ -22,6 +22,7 @@
     <StackPanel
                             Margin="8,6,8,8">
         <StackPanel
+            DataContext="{Binding TransitionsPresentersVM}"
                                 Orientation="Horizontal"
                                 Margin="0,0,0,8">
             <elements:ToggleButtonEx
